+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.6.0 / 2023-04-05
   * Added attribution support
   * Renamed `jekyll_plugin_support_helper.rb` to `jekyll_plugin_helper.rb`.
+  * Now returns `undefined` instead of `false'` when `@keys_values.empty?`
+    and parameter is not specified.
 
 ## 0.5.3 / 2023-03-19
   * Added properties `argv_original`, `keys_values_original`, `remaining_markup_original`,